Firewalls/PaloAlto/requirements.txt
PyYAML
reportlab